+++ coreTests/core.test.ts
@@ -1183,6 +1183,22 @@ c`)
     equal(c.getSiblings().length, 2)
     equal(c.getOlderSiblings().length, 2)
     equal(c.getYoungerSiblings().length, 0)
+
+    // Act
+    a.appendSibling("a2", "foo")
+    a.prependSibling("a-1", "foo")
+
+    // Assert
+    equal(
+      test.toString(),
+      `a-1
+ foo
+a
+a2
+ foo
+b
+c`
+    )
   }
 
   testTree.replaceNode = equal => {