+++ products/treeBase.node.js
@@ -6,6 +6,8 @@
 //tooling product SandboxServer.node.js
 //tooling product core.test.browser.js
 //tooling product abstractBuilder.node.js
+//tooling product TreeComponentFramework.browser.js
+//tooling product TreeComponentFramework.node.js
 Object.defineProperty(exports, "__esModule", { value: true })
 //tooling product treeBase.node.js
 //tooling product abstractBuilder.node.js